RF(BK): use URL to create local file URL -- need to unify encode/decode steps
